ingestor/data_ingestor/requirements.txt
requests
google-cloud-pubsub
google-generativeai
pyyaml 